refactor(scaffold): promote raycastImplicit + directionFromAngles helpers (#164 step 1)
Move the two pure helpers introduced by tangent-planes' #147 PR into
the scaffold layer where they're now consumed by both `tangent-planes`
and `gradient-levels` (incoming in #164 step 2).

Mechanical move + import-path update — no source-code changes to
either helper. Filename rename `raycastSurface.ts` → `raycastImplicit.ts`
aligns with its exported function and matches its sibling
`scaffold/render/ImplicitSurface.ts`.

Adds three test cases to `raycastImplicit.test.ts` covering
`f(origin) = 0` with no downstream sign change (cone-apex scenarios
gradient-levels exercises at k = 0). The strict-less-than check in
`fPrev * fNext < 0` correctly rejects `0 · x` as a sign change, so
miss is returned in all three cases — the new tests lock that.
